refactor(phase3): split src/index.ts monolith into src/ssr/pages/*
Completes Wave 2 of Phase 3 (decomposition). The last and largest of
the three Wave-2 owners.

## src/index.ts: 1,599 -> 398 lines (75% reduction)
Deleted every inline SSR page handler and the local parseContent/
decodeHtmlEntities/toPlainText/escapeHtml/NAMED_ENTITIES duplicates
(now imported from utils/textUtils, matching the rest of Wave 2).
index.ts is now what CLAUDE.md's own conventions always asked for:
imports, app creation, the canonical/static-first middleware, a couple
of small local helpers (serveAsset/serveBucketJson/requestAcceptsBrotli)
still genuinely worker-entry-level, route registrations calling into
the extracted modules, and the catch-all 404. All `c: any` in code
touched during the extraction is now properly `Context<{Bindings:
WorkerEnv}>`.

## New src/ssr/pages/ (4 files + shared)
- shared.ts: cross-page constants/helpers/types (WorkerEnv, AppContext,
  Section, SpeechIndexRow, SpeechListItem, SpeakerListItem,
  AlternateInfo, PAGEFIND_SCRIPT, STATS_SCRIPT, excludedPaths, etc).
- home.ts: renderHomePage, renderSpeechesPage, renderSpeakersPage,
  renderPrivacyPage, renderTermsPage + their loadSpeeches/loadSpeakers/
  buildSpeechListDataToken helpers.
- search.ts: runSearchQuery, renderSearchPage, normalizeSearchQuery +
  the SEARCH_* constants and the single-copy (not in textUtils.ts)
  escapeRegExp/tokenizeSearchQuery/highlightSearchText/
  buildSearchSnippet helpers.
- speech.ts: renderSectionPage (/speech/:section_id),
  renderNestedSpeechPage (/:filename/:nest_filename), renderSpeechPage
  (/:filename) — the three largest extracted handlers — plus
  loadSection/loadSpeechMeta/loadSpeechRedirect/loadAlternateInfo/
  parseToArray.
- speaker.ts: renderSpeakerPage (/speaker/:route_pathname) with typed
  SpeakerRow/SpeakerSectionRow domain types replacing inline `any`.

Pure code motion — behavior preserved exactly, verified by the full
suite passing unchanged (response bodies, status codes, headers, route
paths all untouched).

`/api/purge_cache` and `/api/cleanup_old_cache` were deliberately left
in index.ts as-is: they're bulk administrative purges
(purgeEverything, delete-old-version-keys), not per-content-change
invalidation, so they don't map onto cachePlan.ts's
planListInvalidation/planCacheInvalidation shape — forcing that fit
would have been a worse abstraction than leaving them alone.

## Lint cleanup (caught by bun run check on the integrated result)
Three unused imports the extraction left behind in the new search.ts:
SEARCH_API_CACHE_CONTROL, decodeHtmlEntities, extractDate — none are
referenced by the functions that landed in that file.

## Verification
lint / fmt:check / typecheck / build (both plain and CLOUDFLARE_ENV=
staging) / wrangler deploy --dry-run --env staging (confirmed correct
staging-only bindings) all pass. Full suite: 49 files, 494 tests, 0
failures. Coverage: 100% statements/functions/lines on every single
file in the repo, including every new src/ssr/pages/* file and
sectionPatch.ts/textUtils.ts (both were 0% before Wave 2 wired them
into real callers). Overall branch coverage 90.51%, matching the
project's documented (not gated) threshold. scripts/verify-lsc.mjs
confirms all 4 annotated pure-core files match their recorded
verified/error baselines, no regressions. git status on the separate
~/w/sayit-hono checkout confirmed clean throughout Wave 2 by all three
subagents and by the orchestrator.

This closes Phase 3 (Decompose) of the tech-debt remediation plan.
Next: Phase 4 (test consolidation) and Phase 5 (staging deploy +
smoke test).
