+++ src/index.ts
@@ -1,18 +1,7 @@
 import { Hono, type Context } from 'hono';
 import { speechIndex } from './api/speech_index';
 import { handleOptions } from './api/cors';
-import {
-	CACHE_KEY_VERSION,
-	DEFAULT_HTML_CACHE_CONTROL,
-	SEARCH_API_CACHE_CONTROL,
-	SEARCH_HTML_CACHE_CONTROL,
-	buildR2HtmlKey,
-	purgeWorkersCache,
-	readR2Cache,
-	tags,
-	withCacheHeaders,
-	writeR2Cache,
-} from './api/cache';
+import { CACHE_KEY_VERSION, SEARCH_API_CACHE_CONTROL, purgeWorkersCache, tags, withCacheHeaders } from './api/cache';
 
 import { speakersIndex } from './api/speakers_index';
 import { speakerDetail } from './api/speaker_detail';
@@ -26,37 +15,6 @@ import { isAuthorizedFromHeader } from './api/auth';
 import { rssFeed } from './api/rss';
 import { handleOgImage, handleOgSpeechImage } from './api/og_routes';
 import { ogLoader } from './api/og_loader';
-import type { ApiEnv } from './api/types';
-import HomeView, { styles as HomeViewStyles } from './views/HomeView.vue';
-import SingleParagraphView, { styles as SingleParagraphViewStyles } from './views/SingleParagraphView.vue';
-import SingleSpeechView, { styles as SingleSpeechViewStyles } from './views/SingleSpeechView.vue';
-import NestedSpeechView, { styles as NestedSpeechViewStyles } from './views/NestedSpeechView.vue';
-import SingleNestedSpeechView, { styles as SingleNestedSpeechViewStyles } from './views/SingleNestedSpeechView.vue';
-import SingleSpeakerView, { styles as SingleSpeakerViewStyles } from './views/SingleSpeakerView.vue';
-import SearchResultView, { styles as SearchResultViewStyles } from './views/SearchResultView.vue';
-import SpeechesView, { styles as SpeechesViewStyles } from './views/SpeechesView.vue';
-import SpeakersView, { styles as SpeakersViewStyles } from './views/SpeakersView.vue';
-import LegalPrivacyView, { styles as LegalPrivacyViewStyles } from './views/LegalPrivacyView.vue';
-import LegalTermsView, { styles as LegalTermsViewStyles } from './views/LegalTermsView.vue';
-import Navbar, { styles as NavbarStyles } from './components/Navbar.vue';
-import Footer, { styles as FooterStyles } from './components/Footer.vue';
-import { renderHtml } from './ssr/render';
-import {
-	headForSpeechContent,
-	headForSingleSpeech,
-	headForSpeaker,
-	headForNestedSpeech,
-	headForNestedSpeechDetail,
-	headForSpeeches,
-	headForSearch,
-	headForSpeakers,
-	headForHome,
-	headForPrivacy,
-	headForTerms,
-} from './ssr/heads';
-import { buildPaginationPages } from './utils/pagination';
-import { normalizeSections } from './utils/sectionUtils';
-import { extractDate } from './search/docBuilder';
 import {
 	SEARCH_INDEX_BASELINE_BR_KEY,
 	SEARCH_INDEX_BASELINE_KEY,
@@ -65,54 +23,17 @@ import {
 	SEARCH_UPDATES_PREFIX,
 	createEmptySearchOverlayManifest,
 } from './search/indexFormat';
+import { decodeHtmlEntities } from './utils/textUtils';
+import { extractDate } from './search/docBuilder';
 
-type WorkerEnv = ApiEnv['Bindings'];
+import type { WorkerEnv } from './ssr/pages/shared';
+import { isExcludedPath } from './ssr/pages/shared';
+import { renderHomePage, renderSpeechesPage, renderSpeakersPage, renderPrivacyPage, renderTermsPage } from './ssr/pages/home';
+import { normalizeSearchQuery, runSearchQuery, renderSearchPage, SEARCH_DEFAULT_PAGE_SIZE, SEARCH_MAX_PAGE_SIZE } from './ssr/pages/search';
+import { renderSectionPage, renderNestedSpeechPage, renderSpeechPage } from './ssr/pages/speech';
+import { renderSpeakerPage } from './ssr/pages/speaker';
 
 const app = new Hono<{ Bindings: WorkerEnv }>();
-
-const SEARCH_MIN_QUERY_LENGTH = 2;
-const SEARCH_MAX_QUERY_LENGTH = 80;
-const SEARCH_DEFAULT_PAGE_SIZE = 20;
-const SEARCH_MAX_PAGE_SIZE = 50;
-const SEARCH_SPEAKER_LIMIT = 10;
-const PAGEFIND_SCRIPT = '<script src="/static/speeches/js/pagefind-search.js?v=au-fugu-12"></script>';
-const TWITTER_WIDGETS_SCRIPT = '<script async src="https://platform.twitter.com/widgets.js" charset="utf-8"></script>';
-
-function hasTwitterEmbed(contents: Array<string | null | undefined>): boolean {
-	return contents.some((c) => typeof c === 'string' && c.includes('twitter-tweet'));
-}
-const STATS_SCRIPT = `<script>(function(){fetch('/stats.json').then(function(r){return r.json()}).then(function(s){var fmt=function(n){return n.toString().replace(/\\B(?=(\\d{3})+(?!\\d))/g,',')};var e;e=document.getElementById('sayit-stat-speeches');if(e)e.textContent=fmt(s.speeches);e=document.getElementById('sayit-stat-speakers');if(e)e.textContent=fmt(s.speakers);e=document.getElementById('sayit-stat-sections');if(e)e.textContent=fmt(s.sections)}).catch(function(){})})()</script>`;
-const NAMED_ENTITIES: Record<string, string> = {
-	amp: '&',
-	lt: '<',
-	gt: '>',
-	quot: '"',
-	apos: "'",
-	nbsp: ' ',
-};
-
-const excludedPaths = [
-	'api',
-	'speeches',
-	'speakers',
-	'speaker',
-	'speech',
-	'og',
-	'rss.xml',
-	'feed.xml',
-	'search',
-	'privacy',
-	'terms',
-	'favicon.ico',
-	'robots.txt',
-	'static',
-	'media',
-	'index.html',
-];
-
-function isExcludedPath(segment: string) {
-	return excludedPaths.includes(segment.toLowerCase());
-}
 
 function normalizeSpeakerPageSearch(url: URL): string {
 	const rawPage = url.searchParams.get('page');
@@ -177,7 +98,7 @@ function buildCanonicalPageUrl(requestUrl: string): string | null {
 	return canonicalUrl.toString();
 }
 
-async function canonicalHtmlPageMiddleware(c: any, next: () => Promise<void>) {
+async function canonicalHtmlPageMiddleware(c: Context<{ Bindings: WorkerEnv }>, next: () => Promise<void>) {
 	if (c.req.method !== 'GET' && c.req.method !== 'HEAD') {
 		return next();
 	}
@@ -197,17 +118,13 @@ app.use('*', canonicalHtmlPageMiddleware);
 // 靜態檔優先：先嘗試 ASSETS（Cloudflare CI 建置），找不到再走 API/SSR
 app.use('*', staticFirstMiddleware);
 
-function buildCacheKey(url: string, { includeSearch = true }: { includeSearch?: boolean } = {}): string {
-	return buildR2HtmlKey(url, { includeSearch });
-}
-
-function requestAcceptsBrotli(c: any): boolean {
+function requestAcceptsBrotli(c: Context<{ Bindings: WorkerEnv }>): boolean {
 	const acceptEncoding = c.req.header('Accept-Encoding') ?? '';
 	return /\bbr\b/i.test(acceptEncoding);
 }
 
 async function serveBucketJson(
-	c: any,
+	c: Context<{ Bindings: WorkerEnv }>,
 	key: string,
 	{
 		cacheControl,
@@ -238,14 +155,14 @@ async function serveBucketJson(
 }
 
 /** 向 Cloudflare 靜態資源 (ASSETS) 要求檔案，path 可指定子路徑，未指定則用請求 URL */
-async function serveAsset(c: any, path?: string): Promise<Response> {
+async function serveAsset(c: Context<{ Bindings: WorkerEnv }>, path?: string): Promise<Response> {
 	const url = new URL(c.req.url);
 	const assetUrl = path ? new URL(path, url) : url;
 	return c.env.ASSETS.fetch(assetUrl.toString());
 }
 
 /** 優先嘗試從 ASSETS 回應靜態檔，找不到再交給後續 API/SSR 路由 */
-async function staticFirstMiddleware(c: any, next: () => Promise<void>) {
+async function staticFirstMiddleware(c: Context<{ Bindings: WorkerEnv }>, next: () => Promise<void>) {
 	const pathname = new URL(c.req.url).pathname;
 	if (
 		pathname.startsWith('/api/') ||
@@ -274,512 +191,6 @@ async function staticFirstMiddleware(c: any, next: () => Promise<void>) {
 	if (res && res.status >= 200 && res.status < 400) return res;
 	return next();
 }
-
-function parseContent(raw?: string | null) {
-	if (!raw) return '';
-	try {
-		const parsed = JSON.parse(raw);
-		return typeof parsed === 'string' ? parsed : raw;
-	} catch {
-		return raw;
-	}
-}
-
-function decodeHtmlEntities(value: string) {
-	return value.replace(/&(#x?[0-9a-fA-F]+|[a-zA-Z]+);/g, (match, entity: string) => {
-		if (entity[0] === '#') {
-			const isHex = entity[1]?.toLowerCase() === 'x';
-			const raw = isHex ? entity.slice(2) : entity.slice(1);
-			const parsed = Number.parseInt(raw, isHex ? 16 : 10);
-			if (!Number.isFinite(parsed) || parsed < 0 || parsed > 0x10ffff) {
-				return match;
-			}
-			return String.fromCodePoint(parsed);
-		}
-
-		return NAMED_ENTITIES[entity] ?? match;
-	});
-}
-
-function toPlainText(html: string) {
-	return decodeHtmlEntities(
-		html
-			.replace(/<style\b[\s\S]*?<\/style>/gi, ' ')
-			.replace(/<script\b[\s\S]*?<\/script>/gi, ' ')
-			.replace(/<br\s*\/?>/gi, ' ')
-			.replace(/<\/(p|div|section|article|li|blockquote|h[1-6]|tr|td|th)>/gi, ' ')
-			.replace(/<[^>]+>/g, ' '),
-	)
-		.replace(/\s+/g, ' ')
-		.trim();
-}
-
-function escapeHtml(value: string) {
-	return value.replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;').replace(/"/g, '&quot;').replace(/'/g, '&#39;');
-}
-
-function tokenizeSearchQuery(value: string): string[] {
-	return value
-		.trim()
-		.split(/\s+/)
-		.map((part) => part.trim())
-		.filter(Boolean);
-}
-
-function escapeRegExp(value: string) {
-	return value.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-}
-
-function highlightSearchText(value: string, query: string): string {
-	const tokens = tokenizeSearchQuery(query);
-	if (!value || tokens.length === 0) return escapeHtml(value || '');
-
-	let highlighted = escapeHtml(value);
-	for (const token of tokens) {
-		const pattern = new RegExp(escapeRegExp(escapeHtml(token)), 'gi');
-		highlighted = highlighted.replace(pattern, (match) => `<em>${match}</em>`);
-	}
-	return highlighted;
-}
-
-function parseToArray(raw?: string | null): string[] {
-	if (!raw) return [];
-	return parseContent(raw)
-		.split(',')
-		.map((v) => v.trim())
-		.filter(Boolean);
-}
-
-function buildSearchSnippet(raw: string, query: string, radius = 80): string {
-	const text = raw.replace(/\s+/g, ' ').trim();
-	if (!text) return '';
-
-	const lowerText = text.toLowerCase();
-	const lowerQuery = query.toLowerCase();
-	const matchIndex = lowerText.indexOf(lowerQuery);
-
-	if (matchIndex < 0) {
-		return text.length <= radius * 2 ? text : `${text.slice(0, radius * 2).trimEnd()}...`;
-	}
-
-	const start = Math.max(0, matchIndex - Math.floor(radius / 2));
-	const end = Math.min(text.length, matchIndex + query.length + radius);
-	let snippet = text.slice(start, end).trim();
-	if (start > 0) snippet = `...${snippet}`;
-	if (end < text.length) snippet = `${snippet}...`;
-	return snippet;
-}
-
-async function loadSection(c: any, sectionId: number) {
-	const row = await c.env.DB.prepare(
-		`SELECT
-			a.filename,
-			a.nest_filename,
-			a.nest_display_name,
-			a.section_id,
-			a.previous_section_id,
-			a.next_section_id,
-			a.section_speaker,
-			a.section_content,
-			si.display_name,
-			sp.photoURL,
-			sp.name,
-			prev_section.section_content AS previous_content,
-			next_section.section_content AS next_content
-		FROM speech_content a
-		LEFT JOIN speech_index si ON a.filename = si.filename
-		LEFT JOIN speakers sp ON a.section_speaker = sp.route_pathname
-		LEFT JOIN speech_content prev_section ON a.section_id = prev_section.next_section_id
-		LEFT JOIN speech_content next_section ON a.section_id = next_section.previous_section_id
-		WHERE a.section_id = ?`,
-	)
-		.bind(sectionId)
-		.first();
-	return row as any;
-}
-
-type Section = {
-	filename: string;
-	nest_filename?: string | null;
-	nest_display_name?: string | null;
-	section_id: number;
-	previous_section_id: number | null;
-	next_section_id: number | null;
-	section_speaker: string | null;
-	section_content: string;
-	display_name: string;
-	photoURL: string | null;
-	name: string | null;
-};
-
-type SpeechIndexRow = {
-	filename: string;
-	display_name: string;
-	isNested: number | boolean;
-	nest_filenames?: string | null;
-	nest_display_names?: string | null;
-};
-
-type SpeechListItem = {
-	filename: string;
-	display_name: string;
-};
-
-type SpeakerListItem = {
-	id: number;
-	route_pathname: string;
-	name: string;
-	photoURL: string | null;
-};
-
-type SearchResultRow = {
-	filename: string;
-	nest_filename?: string | null;
-	display_name: string;
-	section_id: number | string | null;
-	section_content: string | null;
-	speaker_name?: string | null;
-};
-
-type SearchSpeakerRow = {
-	id: number;
-	route_pathname: string;
-	name: string;
-	photoURL: string | null;
-};
-
-type SearchSpeakerResult = {
-	route_pathname: string;
-	name: string;
-	photoURL: string | null;
-	snippet: string;
-};
-
-type SearchSectionResult = {
-	section_id: number;
-	filename: string;
-	nest_filename: string | null;
-	section_speaker: string | null;
-	speaker_name: string | null;
-	display_name: string;
-	photoURL: string | null;
-	snippet: string;
-};
-
-type SearchPageResult = {
-	query: string;
-	speakers: SearchSpeakerResult[];
-	sections: SearchSectionResult[];
-	page: number;
-	page_size: number;
-	total_pages: number;
-	total_sections: number;
-	pagination_pages: Array<number | 'ellipsis'>;
-	filteredSpeakerId?: number;
-	filteredSpeakerName?: string | null;
-};
-
-async function loadSpeechMeta(c: any, filename: string): Promise<SpeechIndexRow | null> {
-	const result = await c.env.DB.prepare(
-		`SELECT filename, display_name, isNested, nest_filenames, nest_display_names
-		 FROM speech_index WHERE filename = ?`,
-	)
-		.bind(filename)
-		.first();
-
-	return (result as SpeechIndexRow) ?? null;
-}
-
-/**
- * 查 speech_redirects：若 old_filename 命中，回傳要跳轉到的 new_filename。
- * Route handler 應在 loadSpeechMeta 回 null 時、回 404 之前呼叫，命中則 301 到正規版本。
- */
-async function loadSpeechRedirect(c: any, oldFilename: string): Promise<string | null> {
-	try {
-		const row = await c.env.DB.prepare('SELECT new_filename FROM speech_redirects WHERE old_filename = ?').bind(oldFilename).first();
-		const target = (row as { new_filename?: string } | null)?.new_filename;
-		return typeof target === 'string' && target.length > 0 ? target : null;
-	} catch (err) {
-		console.error('[speech redirect] DB error', err);
-		return null;
-	}
-}
-
-function buildSpeechRedirectResponse(c: any, location: string): Response {
-	const response = c.redirect(location, 301);
-	response.headers.set('Cache-Control', 'public, max-age=86400');
-	return response;
-}
-
-type AlternateInfo = { url: string; label: string; displayName: string; hreflang: string };
-
-async function loadAlternateInfo(c: any, filename: string): Promise<AlternateInfo | null> {
-	try {
-		const row = await c.env.DB.prepare(
-			`SELECT si.alternate_filename, alt.display_name AS alternate_display_name
-			 FROM speech_index si
-			 LEFT JOIN speech_index alt ON si.alternate_filename = alt.filename
-			 WHERE si.filename = ? AND si.alternate_filename IS NOT NULL`,
-		)
-			.bind(filename)
-			.first();
-		if (!row?.alternate_filename) return null;
-		const displayName = row.alternate_display_name || row.alternate_filename;
-		const isCjk = /[\u4e00-\u9fff]/.test(displayName);
-		return {
-			url: `/${encodeURIComponent(row.alternate_filename)}`,
-			label: isCjk ? '華文' : 'English',
-			displayName,
-			hreflang: isCjk ? 'zh-Hant' : 'en',
-		};
-	} catch {
-		return null;
-	}
-}
-
-async function loadSpeeches(c: any): Promise<SpeechListItem[]> {
-	const result = await c.env.DB.prepare('SELECT filename, display_name FROM speech_index ORDER BY id ASC').all();
-
-	if (!result.success) {
-		throw new Error('Database query failed');
-	}
-
-	return result.results.map((row: any) => ({
-		filename: row.filename,
-		display_name: row.display_name,
-	}));
-}
-
-async function buildSpeechListDataToken(speeches: SpeechListItem[]): Promise<string> {
-	const payload = speeches.map((speech) => `${speech.filename}\u0000${speech.display_name}`).join('\u0001');
-	const digest = await crypto.subtle.digest('SHA-256', new TextEncoder().encode(payload));
-	const bytes = Array.from(new Uint8Array(digest.slice(0, 16)));
-	const hash = bytes.map((byte) => byte.toString(16).padStart(2, '0')).join('');
-	return `${speeches.length}-${hash}`;
-}
-
-async function loadSpeakers(c: any): Promise<SpeakerListItem[]> {
-	const result = await c.env.DB.prepare(
-		`SELECT id, route_pathname, name,
-			COALESCE(photoURL, (
-				SELECT s2.photoURL FROM speakers s2
-				WHERE s2.name = speakers.name AND s2.photoURL IS NOT NULL
-				ORDER BY s2.id ASC LIMIT 1
-			)) AS photoURL
-		FROM speakers ORDER BY id ASC`,
-	).all();
-
-	if (!result.success) {
-		throw new Error('Database query failed');
-	}
-
-	return result.results.map((row: any) => ({
-		id: row.id,
-		route_pathname: row.route_pathname,
-		name: row.name,
-		photoURL: row.photoURL ?? null,
-	}));
-}
-
-function normalizeSearchQuery(raw: string | null | undefined): string {
-	const value = (raw || '').trim().slice(0, SEARCH_MAX_QUERY_LENGTH);
-	return value;
-}
-
-async function runSearchQuery(
-	c: any,
-	{
-		query,
-		page = 1,
-		pageSize = SEARCH_DEFAULT_PAGE_SIZE,
-		speakerId,
-	}: {
-		query: string;
-		page?: number;
-		pageSize?: number;
-		speakerId?: number;
-	},
-): Promise<SearchPageResult> {
-	const normalizedQuery = normalizeSearchQuery(query);
-	const safePageSize = Math.max(1, Math.min(SEARCH_MAX_PAGE_SIZE, Math.floor(pageSize || SEARCH_DEFAULT_PAGE_SIZE)));
-	const safePage = Math.max(1, Math.floor(page || 1));
-
-	let filteredSpeakerId: number | undefined;
-	let filteredSpeakerName: string | null = null;
-	let filteredSpeakerRoutePathname: string | null = null;
-	if (speakerId && Number.isFinite(speakerId) && speakerId > 0) {
-		const speakerRow = (await c.env.DB.prepare('SELECT id, route_pathname, name FROM speakers WHERE id = ?')
-			.bind(Math.floor(speakerId))
-			.first()) as { id: number; route_pathname: string; name: string } | null;
-		if (speakerRow?.route_pathname) {
-			filteredSpeakerId = Number(speakerRow.id);
-			filteredSpeakerRoutePathname = speakerRow.route_pathname;
-			filteredSpeakerName = speakerRow.name ?? null;
-		}
-	}
-
-	if (normalizedQuery.length < SEARCH_MIN_QUERY_LENGTH) {
-		return {
-			query: normalizedQuery,
-			speakers: [],
-			sections: [],
-			page: 1,
-			page_size: safePageSize,
-			total_pages: 1,
-			total_sections: 0,
-			pagination_pages: [1],
-			filteredSpeakerId,
-			filteredSpeakerName,
-		};
-	}
-
-	const normalizedLowerQuery = normalizedQuery.toLowerCase();
-	const sectionFilterSql = filteredSpeakerRoutePathname ? 'AND sc.section_speaker = ?' : '';
-	const sectionFilterBindings = filteredSpeakerRoutePathname ? [filteredSpeakerRoutePathname] : [];
-
-	const [speakerResult, totalSectionsRow] = await Promise.all([
-		filteredSpeakerRoutePathname
-			? Promise.resolve({ success: true, results: [] as SearchSpeakerRow[] })
-			: c.env.DB.prepare(
-					`SELECT id, route_pathname, name, photoURL
-				FROM speakers
-				WHERE instr(lower(COALESCE(name, '')), ?) > 0
-				ORDER BY instr(lower(COALESCE(name, '')), ?), name ASC
-				LIMIT ?`,
-				)
-					.bind(normalizedLowerQuery, normalizedLowerQuery, SEARCH_SPEAKER_LIMIT)
-					.all(),
-		c.env.DB.prepare(
-			`SELECT COUNT(*) AS count
-			FROM speech_content sc
-			LEFT JOIN speech_index si ON sc.filename = si.filename
-			LEFT JOIN speakers sp ON sc.section_speaker = sp.route_pathname
-			WHERE (
-				instr(lower(COALESCE(si.display_name, '')), ?) > 0
-				OR instr(lower(COALESCE(sp.name, '')), ?) > 0
-				OR instr(lower(COALESCE(sc.section_content, '')), ?) > 0
-			) ${sectionFilterSql}`,
-		)
-			.bind(normalizedLowerQuery, normalizedLowerQuery, normalizedLowerQuery, ...sectionFilterBindings)
-			.first() as Promise<{ count: number | string | null } | null>,
-	]);
-
-	if (!speakerResult.success) {
-		throw new Error('Database query failed');
-	}
-
-	const totalSections = Number(totalSectionsRow?.count ?? 0);
-	const totalPages = Math.max(1, Math.ceil(totalSections / safePageSize));
-	const resolvedPage = Math.min(safePage, totalPages);
-	const offset = (resolvedPage - 1) * safePageSize;
-
-	const sectionResult = await c.env.DB.prepare(
-		`SELECT
-			sc.filename,
-			sc.nest_filename,
-			sc.section_id,
-			sc.section_speaker,
-			sc.section_content,
-			si.display_name,
-			sp.name AS speaker_name,
-			sp.photoURL
-		FROM speech_content sc
-		LEFT JOIN speech_index si ON sc.filename = si.filename
-		LEFT JOIN speakers sp ON sc.section_speaker = sp.route_pathname
-		WHERE (
-			instr(lower(COALESCE(si.display_name, '')), ?) > 0
-			OR instr(lower(COALESCE(sp.name, '')), ?) > 0
-			OR instr(lower(COALESCE(sc.section_content, '')), ?) > 0
-		) ${sectionFilterSql}
-		ORDER BY
-			CASE
-				WHEN instr(lower(COALESCE(si.display_name, '')), ?) > 0 THEN 0
-				WHEN instr(lower(COALESCE(sp.name, '')), ?) > 0 THEN 1
-				ELSE 2
-			END,
-			sc.filename DESC,
-			sc.section_id ASC
-		LIMIT ? OFFSET ?`,
-	)
-		.bind(
-			normalizedLowerQuery,
-			normalizedLowerQuery,
-			normalizedLowerQuery,
-			...sectionFilterBindings,
-			normalizedLowerQuery,
-			normalizedLowerQuery,
-			safePageSize,
-			offset,
-		)
-		.all();
-
-	if (!sectionResult.success) {
-		throw new Error('Database query failed');
-	}
-
-	return {
-		query: normalizedQuery,
-		speakers: ((speakerResult.results ?? []) as SearchSpeakerRow[]).map((row: SearchSpeakerRow) => ({
-			route_pathname: row.route_pathname,
-			name: row.name,
-			photoURL: row.photoURL ?? null,
-			snippet: highlightSearchText(row.name ?? '', normalizedQuery),
-		})),
-		sections: ((sectionResult.results ?? []) as Array<SearchResultRow & { section_speaker?: string | null; photoURL?: string | null }>).map(
-			(row) => ({
-				section_id: Number(row.section_id),
-				filename: row.filename,
-				nest_filename: row.nest_filename ?? null,
-				section_speaker: row.section_speaker ?? null,
-				speaker_name: row.speaker_name ?? null,
-				display_name: row.display_name,
-				photoURL: row.photoURL ?? null,
-				snippet: highlightSearchText(
-					buildSearchSnippet(toPlainText(parseContent(row.section_content ?? '')), normalizedQuery),
-					normalizedQuery,
-				),
-			}),
-		),
-		page: resolvedPage,
-		page_size: safePageSize,
-		total_pages: totalPages,
-		total_sections: totalSections,
-		pagination_pages: buildPaginationPages(resolvedPage, totalPages),
-		filteredSpeakerId,
-		filteredSpeakerName,
-	};
-}
-
-async function renderSearchPage(c: Context<{ Bindings: WorkerEnv }>) {
-	const url = new URL(c.req.url);
-	const query = normalizeSearchQuery(url.searchParams.get('q'));
-	const pageParam = Number(url.searchParams.get('page') || '1');
-	const page = Number.isFinite(pageParam) && pageParam > 0 ? Math.floor(pageParam) : 1;
-	const speakerParam = Number(url.searchParams.get('p') || '');
-	const speakerId = Number.isFinite(speakerParam) && speakerParam > 0 ? Math.floor(speakerParam) : undefined;
-
-	let result: SearchPageResult;
-	try {
-		result = await runSearchQuery(c, { query, page, pageSize: SEARCH_DEFAULT_PAGE_SIZE, speakerId });
-	} catch (err) {
-		console.error('[search SSR] query failed', err);
-		return c.text('Internal Server Error', 500);
-	}
-
-	const styles = [SearchResultViewStyles, NavbarStyles, FooterStyles].filter(Boolean).join('\n');
-	const head = headForSearch(result.query);
-	const html = await renderHtml(SearchResultView, {
-		head,
-		styles,
-		components: { Navbar, Footer },
-		props: result,
-		scripts: PAGEFIND_SCRIPT,
-	});
-
-	return withCacheHeaders(c.html(html), SEARCH_HTML_CACHE_CONTROL, [tags.listSearch]);
-}
-
-// /、/speeches、/speakers 由 SSR 路由提供，其餘靜態資源由 staticFirstMiddleware 嘗試從 ASSETS 提供
 
 // API CORS preflight
 app.options('/api/*', (c) => handleOptions(c));
@@ -947,107 +358,6 @@ app.post('/api/cleanup_old_cache', async (c) => {
 	return c.json({ deleted, more: false });
 });
 
-async function renderPrivacyPage(c: Context<{ Bindings: WorkerEnv }>) {
-	const styles = [LegalPrivacyViewStyles, NavbarStyles, FooterStyles].filter(Boolean).join('\n');
-	const head = headForPrivacy();
-	const html = await renderHtml(LegalPrivacyView, {
-		head,
-		styles,
-		components: { Navbar, Footer },
-		props: {},
-	});
-	return withCacheHeaders(c.html(html), DEFAULT_HTML_CACHE_CONTROL, [tags.listPrivacy]);
-}
-
-async function renderTermsPage(c: Context<{ Bindings: WorkerEnv }>) {
-	const styles = [LegalTermsViewStyles, NavbarStyles, FooterStyles].filter(Boolean).join('\n');
-	const head = headForTerms();
-	const html = await renderHtml(LegalTermsView, {
-		head,
-		styles,
-		components: { Navbar, Footer },
-		props: {},
-	});
-	return withCacheHeaders(c.html(html), DEFAULT_HTML_CACHE_CONTROL, [tags.listTerms]);
-}
-
-async function renderHomePage(c: Context<{ Bindings: WorkerEnv }>) {
-	const styles = [HomeViewStyles, NavbarStyles, FooterStyles].filter(Boolean).join('\n');
-	const head = headForHome();
-	const html = await renderHtml(HomeView, {
-		head,
-		styles,
-		components: { Navbar, Footer },
-		props: {},
-		scripts: [PAGEFIND_SCRIPT, STATS_SCRIPT].join('\n'),
-	});
-
-	return withCacheHeaders(c.html(html), DEFAULT_HTML_CACHE_CONTROL, [tags.listHome]);
-}
-
-async function renderSpeechesPage(c: Context<{ Bindings: WorkerEnv }>) {
-	let speeches: SpeechListItem[];
-	try {
-		speeches = await loadSpeeches(c);
-	} catch (err) {
-		console.error('[speeches SSR] DB error', err);
-		return c.text('Internal Server Error', 500);
-	}
-
-	const baseCacheKey = buildCacheKey(c.req.url, { includeSearch: false });
-	const dataToken = await buildSpeechListDataToken(speeches);
-	const cacheKey = `${baseCacheKey}data-${dataToken}`;
-	const r2Cached = await readR2Cache(c.env.SPEECH_CACHE, cacheKey);
-	if (r2Cached) return withCacheHeaders(r2Cached, DEFAULT_HTML_CACHE_CONTROL, [tags.listSpeeches]);
-
-	const styles = [SpeechesViewStyles, NavbarStyles, FooterStyles].filter(Boolean).join('\n');
-	const head = headForSpeeches();
-	const html = await renderHtml(SpeechesView, {
-		head,
-		styles,
-		components: { Navbar, Footer },
-		props: { speeches },
-		scripts: PAGEFIND_SCRIPT,
-	});
-
-	let response = c.html(html);
-	response = withCacheHeaders(response, DEFAULT_HTML_CACHE_CONTROL, [tags.listSpeeches]);
-	if (response.ok && response.status < 400) {
-		await writeR2Cache(c.env.SPEECH_CACHE, cacheKey, response.clone());
-	}
-	return response;
-}
-
-async function renderSpeakersPage(c: Context<{ Bindings: WorkerEnv }>) {
-	const cacheKey = buildCacheKey(c.req.url, { includeSearch: false });
-	const r2Cached = await readR2Cache(c.env.SPEECH_CACHE, cacheKey);
-	if (r2Cached) return withCacheHeaders(r2Cached, DEFAULT_HTML_CACHE_CONTROL, [tags.listSpeakers]);
-
-	let speakers: SpeakerListItem[];
-	try {
-		speakers = await loadSpeakers(c);
-	} catch (err) {
-		console.error('[speakers SSR] DB error', err);
-		return c.text('Internal Server Error', 500);
-	}
-
-	const styles = [SpeakersViewStyles, NavbarStyles, FooterStyles].filter(Boolean).join('\n');
-	const head = headForSpeakers();
-	const html = await renderHtml(SpeakersView, {
-		head,
-		styles,
-		components: { Navbar, Footer },
-		props: { speakers },
-	});
-
-	let response = c.html(html);
-	response = withCacheHeaders(response, DEFAULT_HTML_CACHE_CONTROL, [tags.listSpeakers]);
-	if (response.ok && response.status < 400) {
-		await writeR2Cache(c.env.SPEECH_CACHE, cacheKey, response.clone());
-	}
-	return response;
-}
-
 // /speeches → /speeches/, /speakers → /speakers/, /index.html → / are all redirected
 // by the canonical-URL middleware before these handlers run.
 app.get('/privacy', (c) => renderPrivacyPage(c));
@@ -1057,326 +367,15 @@ app.get('/speakers/', (c) => renderSpeakersPage(c));
 app.get('/', (c) => renderHomePage(c));
 
 // /speech/:section_id -> .md/.an 轉專用處理，否則為動態段落頁
-app.on(['GET', 'HEAD'], '/speech/:section_id', async (c) => {
-	const param = c.req.param('section_id');
-	if (param.endsWith('.md')) {
-		console.log('serving md by key', param);
-		return serveMdByKey(c, param);
-	}
-	if (param.endsWith('.an')) {
-		console.log('serving an by key', param);
-		return serveAnByKey(c, param);
-	}
-	// 以下為動態段落頁
-	const sectionId = Number(param);
-	if (!Number.isInteger(sectionId)) {
-		return c.text('Bad Request', 400);
-	}
-
-	const cacheKey = buildCacheKey(c.req.url, { includeSearch: false });
-	const r2Cached = await readR2Cache(c.env.SPEECH_CACHE, cacheKey);
-	// Cache-Tag restored from R2 customMetadata when present (writeR2Cache stores it).
-	if (r2Cached) {
-		const existingTag = r2Cached.headers.get('Cache-Tag');
-		return withCacheHeaders(r2Cached, DEFAULT_HTML_CACHE_CONTROL, existingTag ? existingTag.split(',') : undefined);
-	}
-
-	let section: any;
-	try {
-		section = await loadSection(c, sectionId);
-	} catch (err) {
-		console.error('[speech page] DB error', err);
-		return c.text('Internal Server Error', 500);
-	}
-
-	if (!section) {
-		return c.text('Not Found', 404);
-	}
-
-	const sectionHtml = parseContent(section.section_content ?? '');
-	const plain = toPlainText(sectionHtml);
-	const snippet = plain ? `${plain.slice(0, 80)}${plain.length > 80 ? '...' : ''}` : (section.display_name ?? '');
-	const titleText = snippet ? `”${snippet}”` : 'View Section';
-	const styles = [SingleParagraphViewStyles, NavbarStyles, FooterStyles].filter(Boolean).join('\n');
-	const navigationScript = `<script>(function(){var box=document.getElementById('keyboard-shortcuts');if(!box)return;var prev=box.getAttribute('data-prev-url')||'';var next=box.getAttribute('data-next-url')||'';function editable(el){if(!el)return false;var tag=el.tagName?el.tagName.toLowerCase():'';return tag==='input'||tag==='textarea'||tag==='select'||tag==='option'||el.isContentEditable;}document.addEventListener('keydown',function(e){if(e.metaKey||e.ctrlKey||e.altKey)return;if(editable(document.activeElement))return;if(e.key==='j'){if(next){window.location.href=next;}}else if(e.key==='k'){if(prev){window.location.href=prev;}}});})();</script>`;
-
-	const head = headForSpeechContent(titleText, sectionId, sectionHtml);
-	const twitterScript = hasTwitterEmbed([section.section_content]) ? TWITTER_WIDGETS_SCRIPT : '';
-	const html = await renderHtml(SingleParagraphView, {
-		head,
-		styles,
-		components: { Navbar, Footer },
-		props: { section },
-		scripts: [navigationScript, PAGEFIND_SCRIPT, twitterScript].filter(Boolean).join('\n'),
-	});
-
-	const response = withCacheHeaders(c.html(html), DEFAULT_HTML_CACHE_CONTROL, [tags.speech(section.filename)]);
-	await writeR2Cache(c.env.SPEECH_CACHE, cacheKey, response.clone());
-	return response;
-});
+app.on(['GET', 'HEAD'], '/speech/:section_id', (c) => renderSectionPage(c));
 
 // SSR 講者頁
-app.get('/speaker/:route_pathname', async (c) => {
-	const cacheKey = buildCacheKey(c.req.url);
-	const routePathname = encodeURIComponent(c.req.param('route_pathname'));
-	const r2Cached = await readR2Cache(c.env.SPEECH_CACHE, cacheKey);
-	if (r2Cached) return withCacheHeaders(r2Cached, DEFAULT_HTML_CACHE_CONTROL, [tags.speaker(routePathname)]);
-
-	console.log(routePathname);
-
-	let speaker: any;
-	let sections: Section[];
-	try {
-		// 取得講者基本資料
-		const speakerRow = await c.env.DB.prepare('SELECT * FROM speakers_view WHERE route_pathname = ?').bind(routePathname).first();
-
-		if (!speakerRow) {
-			return c.text('Not Found', 404);
-		}
-
-		const pageSize = 50;
-		const urlObj = new URL(c.req.url);
-		const pageParam = urlObj.searchParams.get('page') ?? '';
-		const pageNumber = Number(pageParam);
-		const page = Number.isFinite(pageNumber) && pageNumber >= 1 ? Math.floor(pageNumber) : 1;
-		const offset = (page - 1) * pageSize;
-
-		// 取得講者的所有段落（分頁）
-		const sectionsResult = await c.env.DB.prepare(
-			`SELECT
-				sc.filename,
-				sc.section_id,
-				sc.previous_section_id,
-				sc.next_section_id,
-				sc.section_speaker,
-				sc.section_content,
-				si.display_name
-			FROM speech_content sc
-			LEFT JOIN speech_index si ON sc.filename = si.filename
-			WHERE sc.section_speaker = ?
-			ORDER BY sc.filename DESC, sc.section_id ASC
-			LIMIT ? OFFSET ?`,
-		)
-			.bind(routePathname, pageSize, offset)
-			.all();
-
-		if (!sectionsResult.success) {
-			throw new Error('Database query failed');
-		}
-
-		sections = normalizeSections(
-			sectionsResult.results.map((row: any) => ({
-				filename: row.filename,
-				display_name: row.display_name,
-				section_id: row.section_id,
-				previous_section_id: row.previous_section_id,
-				next_section_id: row.next_section_id,
-				section_speaker: row.section_speaker,
-				section_content: row.section_content,
-				photoURL: null,
-				name: null,
-			})),
-			false, // 分頁結果，不做重新串接
-		);
-
-		const longestSection = speakerRow.longest_section_id
-			? {
-					section_id: speakerRow.longest_section_id,
-					section_content: speakerRow.longest_section_content || '',
-					section_filename: speakerRow.longest_section_filename || '',
-					section_display_name: speakerRow.longest_section_displayname || '',
-				}
-			: null;
-
-		const totalSections = (typeof speakerRow.sections_count === 'number' ? speakerRow.sections_count : null) ?? sections.length;
-		const totalPages = Math.max(1, Math.ceil(totalSections / pageSize));
-		const paginationPages = buildPaginationPages(page, totalPages);
-
-		speaker = {
-			id: speakerRow.id,
-			route_pathname: speakerRow.route_pathname,
-			name: speakerRow.name,
-			photoURL: speakerRow.photoURL,
-			appearances_count: speakerRow.appearances_count ?? 0,
-			sections_count: totalSections,
-			page,
-			page_size: pageSize,
-			total_pages: totalPages,
-			pagination_pages: paginationPages,
-			sections,
-			longest_section: longestSection,
-		};
-	} catch (err) {
-		console.error('[speaker SSR] DB error', err);
-		return c.text('Internal Server Error', 500);
-	}
-
-	const styles = [SingleSpeakerViewStyles, NavbarStyles, FooterStyles].filter(Boolean).join('\n');
-	const head = headForSpeaker(speaker.route_pathname);
-
-	const twitterScript = hasTwitterEmbed((speaker.sections ?? []).map((s: any) => s.section_content)) ? TWITTER_WIDGETS_SCRIPT : '';
-	const html = await renderHtml(SingleSpeakerView, {
-		head,
-		styles,
-		components: { Navbar, Footer },
-		props: { initialSpeaker: speaker, routePathname: speaker.route_pathname },
-		scripts: ['<script src="/static/speeches/js/masonry.pkgd.min.js"></script>', PAGEFIND_SCRIPT, twitterScript].filter(Boolean).join('\n'),
-	});
-
-	let response = c.html(html);
-	response = withCacheHeaders(response, DEFAULT_HTML_CACHE_CONTROL, [tags.speaker(routePathname)]);
-
-	if (response.ok && response.status < 400) {
-		console.log('writing to R2 cache', cacheKey);
-		await writeR2Cache(c.env.SPEECH_CACHE, cacheKey, response.clone());
-	}
-
-	return response;
-});
+app.get('/speaker/:route_pathname', (c) => renderSpeakerPage(c));
 
 // favicon、robots、/media/*、/static/* 由 staticFirstMiddleware 從 ASSETS 提供
 
 // SSR 巢狀演講內容頁（巢狀子項）
-app.get('/:filename/:nest_filename', async (c) => {
-	const cacheKey = buildCacheKey(c.req.url, { includeSearch: false });
-	const encodedFilename = c.req.param('filename');
-	const encodedNestFilename = c.req.param('nest_filename');
-
-	if (isExcludedPath(encodedFilename)) {
-		return c.text('Not Found', 404);
-	}
-
-	let filename: string;
-	let nestFilename: string;
-	try {
-		filename = decodeURIComponent(encodedFilename);
-		nestFilename = decodeURIComponent(encodedNestFilename);
-	} catch {
-		return c.text('Not Found', 404);
-	}
-
-	const r2Cached = await readR2Cache(c.env.SPEECH_CACHE, cacheKey);
-	if (r2Cached) return withCacheHeaders(r2Cached, DEFAULT_HTML_CACHE_CONTROL, [tags.speech(filename)]);
-
-	let speechMeta: SpeechIndexRow | null;
-	try {
-		speechMeta = await loadSpeechMeta(c, filename);
-	} catch (err) {
-		console.error('[nested speech meta] DB error', err);
-		return c.text('Internal Server Error', 500);
-	}
-
-	if (!speechMeta) {
-		const redirectTo = await loadSpeechRedirect(c, filename);
-		if (redirectTo) {
-			return buildSpeechRedirectResponse(c, `/${encodeURIComponent(redirectTo)}/${encodeURIComponent(nestFilename)}`);
-		}
-		return c.text('Not Found', 404);
-	}
-
-	if (!speechMeta.isNested) {
-		return c.text('Not Found', 404);
-	}
-
-	let sections: Section[];
-	try {
-		const result = await c.env.DB.prepare(
-			`SELECT
-				sc.filename,
-				sc.nest_filename,
-				sc.nest_display_name,
-				sc.section_id,
-				sc.previous_section_id,
-				sc.next_section_id,
-				sc.section_speaker,
-				sc.section_content,
-				si.display_name,
-				sp.photoURL,
-				sp.name
-			FROM speech_content sc
-			LEFT JOIN speech_index si ON sc.filename = si.filename
-			LEFT JOIN speakers sp ON sc.section_speaker = sp.route_pathname
-			WHERE sc.filename = ? AND sc.nest_filename = ?
-			ORDER BY sc.section_id ASC`,
-		)
-			.bind(filename, nestFilename)
-			.all();
-
-		if (!result.success) {
-			throw new Error('Database query failed');
-		}
-
-		const rawSections = result.results.map((row: any) => ({
-			filename: row.filename,
-			nest_filename: row.nest_filename ?? null,
-			nest_display_name: row.nest_display_name ?? row.nest_filename ?? null,
-			section_id: row.section_id,
-			previous_section_id: row.previous_section_id,
-			next_section_id: row.next_section_id,
-			section_speaker: row.section_speaker,
-			section_content: row.section_content,
-			display_name: row.display_name,
-			photoURL: row.photoURL,
-			name: row.name,
-		}));
-
-		if (rawSections.length === 0) {
-			return c.text('Not Found', 404);
-		}
-
-		sections = normalizeSections(rawSections);
-	} catch (err) {
-		console.error('[nested speech detail] DB error', err);
-		return c.text('Internal Server Error', 500);
-	}
-
-	const nestDisplayName = sections[0]?.nest_display_name ?? nestFilename;
-	const speechDisplayName = speechMeta.display_name ?? filename;
-	const nestFilenames = parseToArray(speechMeta.nest_filenames);
-	const nestDisplayNames = parseToArray(speechMeta.nest_display_names);
-	const siblings = nestFilenames.map((nest, idx) => ({
-		nest_filename: nest,
-		nest_display_name: nestDisplayNames[idx] ?? nest,
-	}));
-	const alternate = await loadAlternateInfo(c, filename);
-	const styles = [SingleNestedSpeechViewStyles, NavbarStyles, FooterStyles].filter(Boolean).join('\n');
-	const head = headForNestedSpeechDetail(nestDisplayName, filename);
-	if (alternate) {
-		head.links = [{ rel: 'alternate', href: `https://archive.tw${alternate.url}`, hreflang: alternate.hreflang }];
-	}
-
-	const hasSiblingNav = siblings.length > 0;
-	const navigationScript = hasSiblingNav
-		? `<script>(function(){var prev=document.querySelector('[data-prev-btn]');var next=document.querySelector('[data-next-btn]');function isEditable(el){if(!el)return false;var tag=el.tagName?el.tagName.toLowerCase():'';return tag==='input'||tag==='textarea'||el.isContentEditable;}document.addEventListener('keydown',function(e){if(e.metaKey||e.ctrlKey||e.altKey)return;if(isEditable(document.activeElement))return;if(e.key==='j'&&next&&next.getAttribute('href')){window.location.href=next.getAttribute('href');}if(e.key==='k'&&prev&&prev.getAttribute('href')){window.location.href=prev.getAttribute('href');}});})();</script>`
-		: undefined;
-
-	const twitterScript = hasTwitterEmbed(sections.map((s: any) => s.section_content)) ? TWITTER_WIDGETS_SCRIPT : '';
-	const html = await renderHtml(SingleNestedSpeechView, {
-		head,
-		styles,
-		components: { Navbar, Footer },
-		props: {
-			sections,
-			speechName: filename,
-			nestFilename,
-			displayName: nestDisplayName,
-			speechDisplayName,
-			siblings,
-			alternateUrl: alternate?.url ?? null,
-			alternateLabel: alternate?.label ?? null,
-		},
-		scripts: [navigationScript, PAGEFIND_SCRIPT, twitterScript].filter(Boolean).join('\n'),
-	});
-
-	let response = c.html(html);
-	response = withCacheHeaders(response, DEFAULT_HTML_CACHE_CONTROL, [tags.speech(filename)]);
-
-	if (response.ok && response.status < 400) {
-		await writeR2Cache(c.env.SPEECH_CACHE, cacheKey, response.clone());
-	}
-
-	return response;
-});
+app.get('/:filename/:nest_filename', (c) => renderNestedSpeechPage(c));
 
 // .md 檔案
 app.on(['GET', 'HEAD'], '/:path{[^/]+\\.md}', (c) => serveMdByKey(c, c.req.param('path')));
@@ -1385,212 +384,7 @@ app.on(['GET', 'HEAD'], '/:path{[^/]+\\.md}', (c) => serveMdByKey(c, c.req.param
 app.on(['GET', 'HEAD'], '/:path{[^/]+\\.an}', (c) => serveAnByKey(c, c.req.param('path')));
 
 // SSR 演講頁（單一演講或巢狀清單，直接用 filename 作為路徑；需置於最後的 catch-all 之前）
-app.get('/:filename', async (c) => {
-	const cacheKey = buildCacheKey(c.req.url, { includeSearch: false });
-	const encodedFilename = c.req.param('filename');
-
-	console.log('SSR Single Speech filename', encodedFilename);
-
-	if (isExcludedPath(encodedFilename)) {
-		return c.text('Not Found', 404);
-	}
-
-	// 純數字留給 /speech/:section_id
-	if (/^\d+$/.test(encodedFilename)) {
-		return c.text('Not Found', 404);
-	}
-
-	let filename: string;
-	try {
-		filename = decodeURIComponent(encodedFilename);
-	} catch {
-		return c.text('Not Found', 404);
-	}
-
-	const r2Cached = await readR2Cache(c.env.SPEECH_CACHE, cacheKey);
-	if (r2Cached) {
-		console.log('[r2 cache] hit', cacheKey);
-		return withCacheHeaders(r2Cached, DEFAULT_HTML_CACHE_CONTROL, [tags.speech(filename)]);
-	}
-
-	let speechMeta: SpeechIndexRow | null;
-	try {
-		speechMeta = await loadSpeechMeta(c, filename);
-	} catch (err) {
-		console.error('[speech meta] DB error', err);
-		return c.text('Internal Server Error', 500);
-	}
-
-	if (!speechMeta) {
-		const redirectTo = await loadSpeechRedirect(c, filename);
-		if (redirectTo) {
-			return buildSpeechRedirectResponse(c, `/${encodeURIComponent(redirectTo)}`);
-		}
-		return c.text('Not Found', 404);
-	}
-
-	if (speechMeta.isNested) {
-		let nests: Array<{ nest_filename: string; nest_display_name: string; section_count: number; preview?: string }> = [];
-
-		try {
-			// Aggregation query: count sections and get first section content per nest
-			const result = await c.env.DB.prepare(
-				`SELECT
-					nest_filename,
-					nest_display_name,
-					COUNT(*) AS section_count,
-					MIN(section_id) AS first_section_id
-				FROM speech_content
-				WHERE filename = ? AND nest_filename IS NOT NULL
-				GROUP BY nest_filename, nest_display_name
-				ORDER BY first_section_id ASC`,
-			)
-				.bind(filename)
-				.all();
-
-			if (!result.success) {
-				throw new Error('Database query failed');
-			}
-
-			// Fetch preview content for each nest's first section in one query
-			const firstIds = (result.results as any[]).map((r: any) => r.first_section_id);
-			let previewMap = new Map<number, string>();
-			if (firstIds.length > 0) {
-				const placeholders = firstIds.map(() => '?').join(',');
-				const previewResult = await c.env.DB.prepare(
-					`SELECT section_id, section_content FROM speech_content WHERE section_id IN (${placeholders})`,
-				)
-					.bind(...firstIds)
-					.all();
-				if (previewResult.success) {
-					for (const row of previewResult.results as any[]) {
-						const parsedContent = parseContent(row.section_content ?? '');
-						const plain = toPlainText(parsedContent);
-						if (plain) previewMap.set(row.section_id, plain.slice(0, 80) + (plain.length > 80 ? '...' : ''));
-					}
-				}
-			}
-
-			nests = (result.results as any[]).map((row: any) => ({
-				nest_filename: row.nest_filename,
-				nest_display_name: row.nest_display_name ?? row.nest_filename,
-				section_count: row.section_count,
-				preview: previewMap.get(row.first_section_id),
-			}));
-		} catch (err) {
-			console.error('[nested speech list] DB error', err);
-			return c.text('Internal Server Error', 500);
-		}
-
-		if (nests.length === 0) {
-			return c.text('Not Found', 404);
-		}
-
-		const alternate = await loadAlternateInfo(c, filename);
-		const styles = [NestedSpeechViewStyles, NavbarStyles, FooterStyles].filter(Boolean).join('\n');
-		const head = headForNestedSpeech(speechMeta.display_name ?? filename, filename);
-		if (alternate) {
-			head.links = [{ rel: 'alternate', href: `https://archive.tw${alternate.url}`, hreflang: alternate.hreflang }];
-		}
-
-		const html = await renderHtml(NestedSpeechView, {
-			head,
-			styles,
-			components: { Navbar, Footer },
-			props: {
-				nests,
-				speechName: filename,
-				displayName: speechMeta.display_name ?? filename,
-				alternateUrl: alternate?.url ?? null,
-				alternateLabel: alternate?.label ?? null,
-			},
-			scripts: PAGEFIND_SCRIPT,
-		});
-
-		let response = c.html(html);
-		response = withCacheHeaders(response, DEFAULT_HTML_CACHE_CONTROL, [tags.speech(filename)]);
-
-		if (response.ok && response.status < 400) {
-			await writeR2Cache(c.env.SPEECH_CACHE, cacheKey, response.clone());
-		}
-
-		return response;
-	}
-
-	let sections: Section[];
-	try {
-		// Query speech_content + speakers only (display_name from speechMeta, skip speech_index join)
-		const result = await c.env.DB.prepare(
-			`SELECT
-				sc.filename,
-				sc.section_id,
-				sc.previous_section_id,
-				sc.next_section_id,
-				sc.section_speaker,
-				sc.section_content,
-				sp.photoURL,
-				sp.name
-			FROM speech_content sc
-			LEFT JOIN speakers sp ON sc.section_speaker = sp.route_pathname
-			WHERE sc.filename = ?
-			ORDER BY sc.section_id ASC`,
-		)
-			.bind(filename)
-			.all();
-
-		if (!result.success) {
-			throw new Error('Database query failed');
-		}
-
-		const displayName_ = speechMeta.display_name ?? filename;
-		const rawSections = result.results.map((row: any) => ({
-			filename: row.filename,
-			section_id: row.section_id,
-			previous_section_id: row.previous_section_id,
-			next_section_id: row.next_section_id,
-			section_speaker: row.section_speaker,
-			section_content: row.section_content,
-			display_name: displayName_,
-			photoURL: row.photoURL ?? null,
-			name: row.name ?? null,
-		}));
-
-		if (rawSections.length === 0) {
-			return c.text('Not Found', 404);
-		}
-
-		sections = normalizeSections(rawSections);
-	} catch (err) {
-		console.error('[speech SSR] DB error', err);
-		return c.text('Internal Server Error', 500);
-	}
-
-	const displayName = sections[0]?.display_name ?? speechMeta.display_name ?? filename;
-	const alternate = await loadAlternateInfo(c, filename);
-	const styles = [SingleSpeechViewStyles, NavbarStyles, FooterStyles].filter(Boolean).join('\n');
-	const head = headForSingleSpeech(displayName, filename);
-	if (alternate) {
-		head.links = [{ rel: 'alternate', href: `https://archive.tw${alternate.url}`, hreflang: alternate.hreflang }];
-	}
-
-	const twitterScript = hasTwitterEmbed(sections.map((s) => s.section_content)) ? TWITTER_WIDGETS_SCRIPT : '';
-	const html = await renderHtml(SingleSpeechView, {
-		head,
-		styles,
-		components: { Navbar, Footer },
-		props: { sections, speechName: filename, displayName, alternateUrl: alternate?.url ?? null, alternateLabel: alternate?.label ?? null },
-		scripts: [PAGEFIND_SCRIPT, twitterScript].filter(Boolean).join('\n'),
-	});
-
-	let response = c.html(html);
-	response = withCacheHeaders(response, DEFAULT_HTML_CACHE_CONTROL, [tags.speech(filename)]);
-
-	if (response.ok && response.status < 400) {
-		await writeR2Cache(c.env.SPEECH_CACHE, cacheKey, response.clone());
-	}
-
-	return response;
-});
+app.get('/:filename', (c) => renderSpeechPage(c));
 
 // 其餘請求：靜態已由 middleware 嘗試過，未匹配則 404
 app.get('*', (c) => c.text('Not Found', 404));
